% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpaca_connect.R
\name{alpaca_connect}
\alias{alpaca_connect}
\title{Connect to an alpaca backend}
\usage{
alpaca_connect(mode, api_key, api_secret)
}
\arguments{
\item{mode}{string. 'paper', 'live' or 'data'}

\item{api_key}{string. corresponding alpaca api key}

\item{api_secret}{string. corresponding alpaca api secret key}
}
\value{
a connection object that you will pass to other rblncr functions
}
\description{
Use this function to specify which credentials to use to connect to Alpaca.
}
\examples{
#'\dontrun{
alpaca_connect(mode = "paper",
               api_key = "REDACTED",
               api_secret = "REDACTED")
}                
}
