% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooldown_elapsed.R
\name{cooldown_elapsed}
\alias{cooldown_elapsed}
\title{Has the cooldown period for a portfolio elapsed?}
\usage{
cooldown_elapsed(last_rebalance, cooldown_days)
}
\arguments{
\item{last_rebalance}{a POSIXct timestamp of the last successful rebalance activity}

\item{cooldown_days}{number of cooldown days post a successful rebalance specified in the portfolio model}
}
\value{
TRUE/FALSE
}
\description{
Has the cooldown period for a portfolio elapsed?
}
\examples{
cooldown_elapsed(lubridate::now() - lubridate::days(30), 35)
}
