% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_portfolio_model.R
\name{create_portfolio_model}
\alias{create_portfolio_model}
\title{Create a portfolio model}
\usage{
create_portfolio_model(name, description, cash, assets, tolerance, cooldown)
}
\arguments{
\item{name}{name of the model}

\item{description}{description of the model}

\item{cash}{cash data frame}

\item{assets}{assets data frame}

\item{tolerance}{percentage by which and symbols is permitted to drift from target weight}

\item{cooldown}{number of days after successful rebalance to wait before new rebalance can be triggered}
}
\value{
a portfolio model object
}
\description{
A portfolio model specifies the target weights of a portfolio, along with drift tolerances and cooldown periods.
}
\examples{
name <- "sample_portfolio"
description <- "create from function"
cash <- list(percent = 10)
assets <- data.frame(symbol = c("AAPL","GOOG"), percent = c(80.5,9.5))
tolerance <- list(percent = 5)
cooldown <- list(days = 365)

create_portfolio_model(name = name,
                       description = description,
                       cash = cash,
                       assets = assets,
                       tolerance = tolerance,
                       cooldown = cooldown)
}
