% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cash.R
\name{get_cash}
\alias{get_cash}
\title{Get cash balance from backend}
\usage{
get_cash(connection)
}
\arguments{
\item{connection}{a trading backend connection}
}
\value{
a data frame cash balance
}
\description{
Get cash balance from backend
}
