% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orders.R
\name{get_orders}
\alias{get_orders}
\title{Get order history from backend}
\usage{
get_orders(connection)
}
\arguments{
\item{connection}{a trading backend connection}
}
\value{
a data frame of orders
}
\description{
Get order history from backend
}
