% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio_current.R
\name{get_portfolio_current}
\alias{get_portfolio_current}
\title{Get current portfolio holdings}
\usage{
get_portfolio_current(connection)
}
\arguments{
\item{connection}{a backend trading connection}
}
\value{
a portfolio object which can be passed to \code{load_portfolio_targets()}
}
\description{
Get current portfolio holdings
}
\examples{
t_conn <- alpaca_connect('paper',
                         Sys.getenv("ALPACA_PAPER_KEY"),
                         Sys.getenv("ALPACA_PAPER_SECRET"))
get_portfolio_current(t_conn)
}
