% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions.R
\name{get_positions}
\alias{get_positions}
\title{Get positions from backend}
\usage{
get_positions(connection)
}
\arguments{
\item{connection}{a trading backend connection}
}
\value{
a data frame of stock positions
}
\description{
Get positions from backend
}
