% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quotes.R
\name{get_quotes}
\alias{get_quotes}
\title{Get quotes from a data backend}
\usage{
get_quotes(symbols, connection)
}
\arguments{
\item{symbols}{a vector of stock symbols}

\item{connection}{a data backend}
}
\value{
a dataframe of quotes
}
\description{
Get quotes from a data backend
}
