% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_portfolio_targets.R
\name{load_portfolio_targets}
\alias{load_portfolio_targets}
\title{Load portfolio targets}
\usage{
load_portfolio_targets(portfolio, portfolio_model)
}
\arguments{
\item{portfolio}{a portfolio object, as returned from \code{get_portfolio_current()}}

\item{portfolio_model}{a portfolio model, as returned from \code{read_portfolio_model()} or \code{create_portfolio_model()}}
}
\value{
an extended portfolio object which can be passed to \code{price_portfolio()}
}
\description{
Load portfolio targets
}
\examples{
t_conn <- alpaca_connect('paper',
                         Sys.getenv("ALPACA_PAPER_KEY"),
                         Sys.getenv("ALPACA_PAPER_SECRET"))

portfolio_model <- read_portfolio_model(system.file(package='rblncr','extdata/sample_portfolio.yaml'))

get_portfolio_current(t_conn) |>
  load_portfolio_targets(portfolio_model)

}
