% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_open.R
\name{market_open}
\alias{market_open}
\title{Determine if the market is open}
\usage{
market_open(connection)
}
\arguments{
\item{connection}{a trading backend connection}
}
\value{
TRUE or FALSE
}
\description{
Determine if the market is open
}
