% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_portfolio_model.R
\name{read_portfolio_model}
\alias{read_portfolio_model}
\title{Read a portfolio model from disk}
\usage{
read_portfolio_model(file_path)
}
\arguments{
\item{file_path}{a file path string}
}
\value{
a portfolio model object
}
\description{
Reads a portfolio model yaml file and returns a portoflio model object
}
\examples{
read_portfolio_model(system.file(package='rblncr','extdata/sample_portfolio.yaml'))

}
