% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_portfolio_model.R
\name{save_portfolio_model}
\alias{save_portfolio_model}
\title{Save a portfolio model to disk}
\usage{
save_portfolio_model(portfolio_model, file_path)
}
\arguments{
\item{portfolio_model}{a portfolio model created with \code{create_portfolio_model()}}

\item{file_path}{a file path string}
}
\value{
file_path
}
\description{
Save a portfolio model to disk
}
