% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_portfolio.R
\name{solve_portfolio}
\alias{solve_portfolio}
\title{Compute which changes are required to make a current portfolio identical to a portfolio model}
\usage{
solve_portfolio(portfolio_priced, terse = TRUE)
}
\arguments{
\item{portfolio_priced}{a portfolio object, as returned from \code{price_portfolio()}}

\item{terse}{TRUE/FALSE should the returned object only contained the information necessary for downstream operations?}
}
\value{
a portfolio object which can be passed to \code{constrain_orders()}
}
\description{
Compute which changes are required to make a current portfolio identical to a portfolio model
}
\examples{
\dontrun{
t_conn <- alpaca_connect('paper',
                         Sys.getenv("ALPACA_PAPER_KEY"),
                         Sys.getenv("ALPACA_PAPER_SECRET"))
d_conn <- alpaca_connect('data',
                         Sys.getenv("ALPACA_LIVE_KEY"),
                         Sys.getenv("ALPACA_LIVE_SECRET"))

portfolio_model <- read_portfolio_model(system.file(package='rblncr','extdata/sample_portfolio.yaml'))

get_portfolio_current(t_conn) |>
  load_portfolio_targets(portfolio_model) |>
  price_portfolio(connection = d_conn, price_type = 'close') |>
  solve_portfolio()
}

}
