% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_portfolio_model.R
\name{validate_portfolio_model}
\alias{validate_portfolio_model}
\title{Validate a portfolio model}
\usage{
validate_portfolio_model(portfolio_model)
}
\arguments{
\item{portfolio_model}{a portfolio model}
}
\value{
TRUE if passed
}
\description{
Validate a portfolio model
}
\examples{
model <- read_portfolio_model(system.file(package='rblncr','extdata/sample_portfolio.yaml'))
validate_portfolio_model(model)
}
